test-requirements.txt
coverage==7.2.7
Faker==18.11.2
pytest==7.4.0
pytest-cov==4.1.0
pytest-django==4.5.2
ruff==0.0.275